requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.1.0
dnspython==2.4.2
idna==3.4
blinker==1.6.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
mysql-connector-python==8.1.0
Flask-Login==0.6.3
PyJWT==2.8.0